requirements.txt
flask==3.0.1
boto3==1.34.23
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.25
pyjwt==2.8.0
psycopg2-binary==2.9.9
flask-marshmallow==1.1.0
python-dotenv==1.0.0
flask-migrate==4.0.5
flask-restful==0.3.10
sqlalchemy-utils==0.41.1
marshmallow-sqlalchemy==0.30.0